+++ src/components/Game/Game.vue
@@ -1,0 +1,15 @@
+<template>
+    <div id="game">
+
+    </div>
+</template>
+
+<script>
+export default {
+    name: "Game",
+};
+</script>
+
+<style scoped>
+
+</style>