+++ src/components/NavBar/NavBar.vue
@@ -9,29 +9,34 @@
 
 <script>
 import Swal from "sweetalert2";
-import { inGameSection } from "../../helpers/functions/Router";
+// import { inGameSection } from "../../helpers/functions/Router";
 
 export default {
     name: "NavBar",
     methods: {
         confirmBackToPage() {
+            const routeNameText = {
+                GameLobby: this.$t("NavBar.gameCompositionNotFinished"),
+                Game: this.$t("NavBar.yourProgressIsSaved"),
+            };
             return Swal.fire({
-                title: "Are you sure?",
-                text: "You won't be able to revert this!",
+                title: this.$t("NavBar.areYouSureYouWantToQuit"),
+                text: routeNameText[this.$route.name],
                 icon: "warning",
                 showCancelButton: true,
-                confirmButtonText: "Yes, delete it!",
+                confirmButtonText: this.$t("NavBar.quit"),
+                cancelButtonText: this.$t("NavBar.cancel"),
             });
         },
         async backToHomePage() {
-            if (inGameSection(this.$route.name)) {
-                const { value } = await this.confirmBackToPage();
-                if (value) {
-                    await this.$router.push("/");
-                }
-            } else {
-                await this.$router.push("/");
-            }
+            // if (inGameSection(this.$route.name)) {
+            //     const { value } = await this.confirmBackToPage();
+            //     if (value) {
+            //         await this.$router.push("/");
+            //     }
+            // } else {
+            // }
+            await this.$router.push("/");
         },
     },
 };