+++ src/plugins/router/index.js
@@ -2,6 +2,7 @@ import Vue from "vue";
 import VueRouter from "vue-router";
 import Home from "../../components/Home/Home";
 import GameLobby from "../../components/GameLobby/GameLobby";
+import Game from "../../components/Game/Game";
 
 Vue.use(VueRouter);
 
@@ -13,6 +14,10 @@ const routes = [{
     path: "/game-lobby",
     name: "GameLobby",
     component: GameLobby,
+}, {
+    path: "/game/:id",
+    name: "Game",
+    component: Game,
 }];
 
 const Router = new VueRouter({